+++ src/main/java/ea/badge/service/BadgeService.java
@@ -6,11 +6,11 @@ import java.util.Collection;
 
 public interface BadgeService {
     Collection<Badge> findAll();
-    Badge findById(Integer id);
-    Collection<Badge> findByMemberId(Integer id);
+    Badge findById(Long id);
+    Collection<Badge> findByMemberId(Long id);
     Badge createOrUpdate(Badge badge);
     Badge replaceWithNew(Badge newBadge);
-    Badge deactivateById(Integer id);
-    Badge activateById(Integer id);
-    void deleteById(Integer id);
+    Badge deactivateById(Long id);
+    Badge activateById(Long id);
+    void deleteById(Long id);
 }
